Added event flag reading - not complete yet but it works for boss flags. Adding autosplitting for bosses only Setup a rust target building a DLL. Can automatically assemble intel based assembly and link it to the DLL, can use detours-rs to detour functions Can do pattern scans while running from inside the games address space, without going through the win32 api.

+++ src/rust_memory/src/memory/pattern_scanner.rs
@@ -1,4 +1,4 @@
-pub fn scan(code: &[u8], pattern: &[Option<u8>]) -> Option<i64>
+pub fn scan(code: &[u8], pattern: &[Option<u8>]) -> Option<usize>
 {
     if code.len() == 0
     {
@@ -21,7 +21,7 @@ pub fn scan(code: &[u8], pattern: &[Option<u8>]) -> Option<i64>
         }
         if found
         {
-            return Some(i as i64);
+            return Some(i);
         }
     }
     return None;
